requirements.txt
chromadb==1.4.1
config==0.5.1
fpdf==1.7.2
joblib==1.5.3
numpy==2.4.1
openai==2.16.0
openai_whisper==20250625
pandas==3.0.0
python-dotenv==1.2.1
reportlab==4.4.9
Requests==2.32.5
scikit_learn==1.8.0
streamlit==1.51.0
yt_dlp==2026.1.29
